cyclecmd: v0.2.0 release content
[FEATURES]:
Added ControlEvents as an additional return type for the Event handler function in order for the user to be able to influence the control flow of the application, e.g.
terminating it.

[ENHANCEMENTS]:
The event loop captured only 1 byte before, this is enough for characters but not enough for escape sequences that are e.g. 3 bytes long. Thus, the event loop
is now able to capture 3 bytes but this introduced timing issues that needs to be resolved in the future by separating parsing from reading data.

[TESTS]:
Needed to change e2e test since user input was simulated by redirecting standard input to read from a file but the buffer will always fill when enough bytes are available but this
does not reflect terminal user input behavior, so this was exchanged with os.Pipe and a time delay between each input.
Removed test cases that checked the validation function that was removed.

Signed-off-by: RaphSku <rapsku.dev@gmail.com>